Moved FollowRequestListAdapter and UserListAdapter into ui folder
